댓글 작성자 정보에 이름 추가
- 댓글 작성자 정보에 이름 추가
- 리팩토링한 코드에 문제가 발생하여 추석처리 후 원상 복구
- 하지만 다른 버그 발생으로 주석 추가

[#2]

+++ pages/api/comment/delete.js
@@ -11,13 +11,25 @@ export default async function handler(request, response) {
 
       const client = await connectDB;
       const db = client.db("forum");
+
+      const findedComment = await db
+        .collection("comment")
+        .findOne({ _id: new ObjectId(request.query.id) });
+
+      // 관리자, 작성자가 아닌 경우 삭제 거부
+      if (
+        session.user.role != "admin" &&
+        session.user.email != findedComment.author.email
+      ) {
+        return response.status(403).json();
+      }
+
       const result = await db.collection("comment").deleteOne({
         _id: new ObjectId(request.query.id),
       });
 
       return response.status(200).json();
     } catch (error) {
-      console.log(error);
       return response.status(500).json();
     }
   }
